Implement rs.utils.is_polar() and rename utils.symop to utils.symmetry
